extract the new graph to a seperate file
